+++ src/extension.ts
@@ -3,6 +3,7 @@ import * as files from './modules/files';
 import * as workspaces from './modules/workspaces';
 import * as folders from './modules/folders';
 import * as crud from './modules/crud';
+import * as pathStrings from './modules/pathStrings';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
@@ -26,6 +27,10 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// workspaces
 	context.subscriptions.push(vscode.commands.registerTextEditorCommand('vscode-FileSystemToolbox.CopyWorkspaceRootPath', () => { workspaces.getWorkspaceRootPath(); }));
+
+	// path transformation
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('vscode-FileSystemToolbox.TransformPathToPosix', () => { pathStrings.transformPath(pathStrings.PathTransformationType.posix); }));
+	context.subscriptions.push(vscode.commands.registerTextEditorCommand('vscode-FileSystemToolbox.TransformPathToWin32', () => { pathStrings.transformPath(pathStrings.PathTransformationType.win32); }));
 }
 
 // this method is called when your extension is deactivated